+++ app/src/main/java/com/pluralmakes/filamentlibrary/MainActivity.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.Modifier
 import com.pluralmakes.filamentlibrary.ui.Collector
 import com.pluralmakes.filamentlibrary.ui.theme.FilamentLibraryTheme
 import com.pluralmakes.filamentlibrary.util.impl.TD1CommunicatorImpl
-import com.pluralmakes.filamentlibrary.viewModel.CollectorViewModel
+import com.pluralmakes.filamentlibrary.model.CollectorViewModel
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
